workshop3/a2a-demo/orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.0.1"
description = "Orchestrator for the A2A demo"
requires-python = ">=3.11"
dependencies = [
  "a2a-sdk>=0.3.3",
  "fastmcp>=2.9.2",
  "uvicorn>=0.34.0",
  "google-adk>=1.9.0",
  "google-genai>=1.17.0",
  "python-dotenv>=1.1.0",
  "fastapi>=0.105.0",
]
